+++ src/nestjs/src/modules/database/userChatRoom/service.ts
@@ -1,4 +1,4 @@
-import { Injectable } from "@nestjs/common";
+import { HttpException, HttpStatus, Injectable } from "@nestjs/common";
 import { Repository } from "typeorm";
 import { InjectRepository } from "@nestjs/typeorm";
 
@@ -21,11 +21,18 @@ export class UserChatRoomService {
 
         async create(post: UserChatRoomPost, UserId: number, RoomId: number) {
                 const userChat = new UserChatRoomEntity();
-                userChat.userId = UserId;
-                userChat.roomId = RoomId;
+                let user = await this.userRepo.findOneBy({id: UserId});
+                if (user)
+                        userChat.userId = UserId;
+                else
+                        throw new HttpException('User not found', HttpStatus.NOT_FOUND);
+                let room = await this.ChatRoomRepo.findOneBy({id: RoomId});
+                if (room)
+                        userChat.roomId = RoomId;
+                else
+                        throw new HttpException('ChatRoom not found', HttpStatus.NOT_FOUND);
                 userChat.isOwner = post.isOwner;
-                // userChat.isAdmin = post.isAdmin;
-                await this.userChatRoomRepo.save(userChat);
+                userChat.isAdmin = post.isAdmin;
                 return await this.userChatRoomRepo.save(userChat);
           }
 
@@ -34,23 +41,23 @@ export class UserChatRoomService {
         }
 
         async returnOne(user: number, room: number) {
-                console.log(user);
-                console.log(room);
-                return await this.userChatRoomRepo.findOneBy({ userId: user, roomId: room });
+                const tmp = await this.userChatRoomRepo.findOneBy({ userId: user, roomId: room });
+                if (tmp)
+                        return tmp;
+                throw new HttpException('Not found', HttpStatus.NOT_FOUND);
         }
 
         async update(user: number, room: number, post: UserChatRoomPost) {
-                console.log(user);
-                console.log(room);
-                console.log(post);
                 const tmp = await this.userChatRoomRepo.findOneBy({ userId: user, roomId: room });
-                return await this.userChatRoomRepo.update(tmp, post);
+                if (tmp)
+                        return await this.userChatRoomRepo.update(tmp, post);
+                throw new HttpException('Not found', HttpStatus.NOT_FOUND);
         }
 
         async delete(user: number, room: number) {
-                console.log(user);
-                console.log(room);
                 const tmp = await this.userChatRoomRepo.findOneBy({ userId: user, roomId: room });
-                return await this.userChatRoomRepo.delete(tmp);
+                if (tmp)
+                        return await this.userChatRoomRepo.delete(tmp);
+                throw new HttpException('Not found', HttpStatus.NOT_FOUND);
         }
 }
